Disomic and polysomic segregation patterns

Gamete frequencies for all possible disomic and polysomic segregation patterns for even ploidies 2 through 20. If you need higher ploidy levels, let me know and I'll update it (it's very easy).

Usage

seg

Format

A data frame with the following columns

ploidy

The ploidy of the parent.

g

The genotype of the parent.

m

The pairing configuration given disomic inheritance. See Gerard et al (2018).

p

The gamete frequencies. Element p[[i]] is the probability a gamete will have dosage i-1.

mode

Whether the inheritance pattern that leads to these gamete frequencies is "disomic", "polysomic", or "both".

References

Gerard, D., Ferrão, L. F. V., Garcia, A. A. F., & Stephens, M. (2018). Genotyping polyploids from messy sequencing data. Genetics, 210(3), 789-807. doi:10.1534/genetics.118.301468

Author

David Gerard