segtest: Tests for Segregation Distortion in Polyploids

Provides tests for segregation distortion in F1 polyploid populations under different assumptions of meiosis. These tests can account for double reduction, partial preferential pairing, and genotype uncertainty through the use of genotype likelihoods. Parallelization support is provided. Details of these methods are described in Gerard et al. (2025) doi:10.1007/s00122-025-04816-z . Part of this material is based upon work supported by the National Science Foundation under Grant No. 2132247. The opinions, findings, and conclusions or recommendations expressed are those of the author and do not necessarily reflect the views of the National Science Foundation.

See also

Author

Maintainer: David Gerard gerard.1787@gmail.com (ORCID)

Authors:

Mira Thakkar chrtmira@gmail.com

Other contributors:

Guilherme da Silva Pereira g.pereira@ufv.br (ORCID) [contributor]

NSF DBI 2132247 (https://www.nsf.gov/awardsearch/showAward?AWD_ID=2132247) [funder]