Parallelized likelihood ratio test for segregation distortion.

Uses the future package to implement parallelization support for the likelihood ratio tests for segregation distortion. This function only works for tetraploids, and cannot account for outliers. For higher ploidies and more functionality, see seg_multi().

Usage

multi_lrt(
  g,
  p1,
  p2,
  drbound = 1/6,
  pp = TRUE,
  dr = TRUE,
  alpha = 0,
  xi1 = 1/3,
  xi2 = 1/3,
  nullprop = FALSE
)

Arguments

g

One of two inputs

A matrix of genotype counts. The rows index the loci and the columns index the genotypes.

An array of genotype log-likelihoods. The rows index the loci, the columns index the individuals, and the slices index the genotypes. Log-likelihoods are base e (natural log).

p1

One of three inputs

A vector of parent 1's genotypes.

A matrix of parent 1's genotype log-likelihoods. The rows index the loci and the columns index the genotypes. Logs are in base e (natural log).

NULL (only supported when using genotype likelihoods for the offspring)

p2

One of three inputs

A vector of parent 1's genotypes.

A matrix of parent 1's genotype log-likelihoods. The rows index the loci and the columns index the genotypes. Logs are in base e (natural log).

NULL (only supported when using genotype likelihoods for the offspring)

drbound

The upper bound on the double reduction rate.

pp

Is (partial) preferential pairing possible (TRUE) or not (FALSE)?

dr

Is double reduction possible (TRUE) or not (FALSE)?

alpha

If dr = FALSE, this is the known rate of double reduction.

xi1

If pp = FALSE, this is the known preferential pairing parameter of parent 1.

xi2

If pp = FALSE, this is the known preferential pairing parameter of parent 2.

nullprop

Should we return the null proportions (TRUE) or not (FALSE)?

Value

A data frame with the following elements:

statistic

The likelihood ratio test statistic

p_value

The p-value of the likelihood ratio test.

df

The degrees of freedom of the test.

alpha

The MLE of the double reduction rate. Do not use for real work.

xi1

The MLE of the first parent's partial preferential pairing parameter. Do not use for real work.

xi2

The MLE of the second parent's partial preferential pairing parameter. Do not use for real work.

p1

(Estimate of) the first parent's genotype.

p2

(Estimate of) the second parent's genotype.

snp

The name of the SNP.

Parallel Computation

The multi_lrt() function supports parallel computing. It does so through the future package.

You first specify the evaluation plan with plan() from the future package. On a local machine, this is typically just future::plan(future::multisession, workers = nc) where nc is the number of workers you want. You can find the maximum number of possible workers with availableCores(). You then run multi_lrt(), then shut down the workers with future::plan(future::sequential).

See also

Author

David Gerard

Examples

# \donttest{
## Assuming genotypes are known (typically a bad idea)
glist <- multidog_to_g(
  mout = ufit,
  ploidy = 4,
  type = "all_g",
  p1 = "indigocrisp",
  p2 = "sweetcrisp")
p1_1 <- glist$p1
p2_1 <- glist$p2
g_1 <- glist$g
multi_lrt(g = g_1, p1 = p1_1, p2 = p2_1)
#> Loading required package: foreach
#> Loading required package: rngtools
#>    statistic     p_value df         alpha       xi1      xi2 p1 p2         snp
#> 1  3.1603821 0.205935746  2 -2.785230e-18        NA 0.433339  1  2  1_44673475
#> 2  1.6700394 0.196253258  1  5.536457e-08        NA       NA  1  0 11_28836119
#> 3  6.8392600 0.008917577  1  9.284955e-02 0.2860801       NA  2  0 12_31029646
#> 4  0.0000000 1.000000000  0            NA        NA       NA  4  4  12_8487773
#> 5  1.0010910 0.800987973  3  2.860567e-02        NA       NA  1  1   2_1426329
#> 6        Inf 0.000000000 NA            NA        NA       NA  0  2  2_20070837
#> 7  1.9586129 0.581041024  3  8.843781e-02 0.6766071       NA  2  3  2_40108108
#> 8  1.9970199 0.157608821  1  1.195716e-02 0.3073001       NA  2  0  4_37820371
#> 9  0.2681554 0.604571837  1  5.536457e-08        NA       NA  0  1  6_30619745
#> 10 0.7975415 0.850054981  3  1.441684e-02        NA       NA  1  1   6_4350249

## Using genotype likelihoods (typically a good idea)
glist <- multidog_to_g(
  mout = ufit,
  ploidy = 4,
  type = "all_gl",
  p1 = "indigocrisp",
  p2 = "sweetcrisp")
p1_2 <- glist$p1
p2_2 <- glist$p2
g_2 <- glist$g
multi_lrt(g = g_2, p1 = p1_2, p2 = p2_2)
#>        statistic     p_value df        alpha       xi1       xi2 p1 p2
#> 1   3.167045e+00 0.205250847  2 0.000000e+00        NA 0.4166706  1  2
#> 2   1.641337e+00 0.200142085  1 5.536457e-08        NA        NA  1  0
#> 3   5.595326e+00 0.018008460  1 5.892102e-02 0.2718967        NA  2  0
#> 4  -1.398643e-07 1.000000000  1 1.666667e-01        NA        NA  4  4
#> 5   1.579334e+00 0.453995964  2 2.055025e-02        NA        NA  1  1
#> 6   1.356856e+01 0.001131421  2 7.198651e-02        NA 0.4621938  0  2
#> 7   2.878633e+00 0.410718596  3 9.343541e-02 0.6564488        NA  2  3
#> 8   1.554209e+00 0.212515165  1 6.190609e-02 0.3632006        NA  2  0
#> 9   2.500730e-01 0.617023681  1 5.536457e-08        NA        NA  0  1
#> 10  8.630885e-01 0.649505305  2 8.077581e-09        NA        NA  1  1
#>            snp
#> 1   1_44673475
#> 2  11_28836119
#> 3  12_31029646
#> 4   12_8487773
#> 5    2_1426329
#> 6   2_20070837
#> 7   2_40108108
#> 8   4_37820371
#> 9   6_30619745
#> 10   6_4350249

## Offspring genotype likelihoods and parent genotypes known
multi_lrt(g = g_2, p1 = p1_1, p2 = p2_1)
#>        statistic     p_value df        alpha       xi1       xi2 p1 p2
#> 1   3.167045e+00 0.205250847  2 0.000000e+00        NA 0.4166706  1  2
#> 2   1.641337e+00 0.200142085  1 5.536457e-08        NA        NA  1  0
#> 3   5.595326e+00 0.018008460  1 6.567195e-02 0.2810866        NA  2  0
#> 4  -1.398643e-07 1.000000000  1 1.666667e-01        NA        NA  4  4
#> 5   1.579334e+00 0.453995964  2 2.055025e-02        NA        NA  1  1
#> 6   1.356856e+01 0.001131421  2 9.423113e-02        NA 0.4981036  0  2
#> 7   2.878633e+00 0.410718596  3 9.343530e-02 0.6564492        NA  2  3
#> 8   1.554209e+00 0.212515165  1 4.085119e-02 0.3332762        NA  2  0
#> 9   2.500730e-01 0.617023681  1 5.536457e-08        NA        NA  0  1
#> 10  8.630885e-01 0.649505305  2 8.077581e-09        NA        NA  1  1
#>            snp
#> 1   1_44673475
#> 2  11_28836119
#> 3  12_31029646
#> 4   12_8487773
#> 5    2_1426329
#> 6   2_20070837
#> 7   2_40108108
#> 8   4_37820371
#> 9   6_30619745
#> 10   6_4350249

## Offspring genotype likelihoods and no information on parent genotypes
multi_lrt(g = g_2, p1 = NULL, p2 = NULL)
#>        statistic     p_value df        alpha       xi1       xi2 p1 p2
#> 1   3.167045e+00 0.205250847  2 0.000000e+00        NA 0.4166705  1  2
#> 2   1.641337e+00 0.200142085  1 5.536457e-08        NA        NA  1  0
#> 3   5.595326e+00 0.018008460  1 2.131367e-02        NA 0.2230223  0  2
#> 4  -1.398643e-07 1.000000000  1 1.666667e-01        NA        NA  4  4
#> 5   1.579334e+00 0.453995964  2 2.055025e-02        NA        NA  1  1
#> 6   1.356856e+01 0.001131421  2 3.475555e-02 0.4057949        NA  2  0
#> 7   2.878633e+00 0.410718596  3 9.343538e-02 0.6564489        NA  2  3
#> 8   1.554209e+00 0.212515165  1 6.576491e-02 0.3688313        NA  2  0
#> 9   2.500730e-01 0.617023681  1 5.536457e-08        NA        NA  1  0
#> 10  8.627833e-01 0.352960981  1 1.791445e-03 0.0029911        NA  2  0
#>            snp
#> 1   1_44673475
#> 2  11_28836119
#> 3  12_31029646
#> 4   12_8487773
#> 5    2_1426329
#> 6   2_20070837
#> 7   2_40108108
#> 8   4_37820371
#> 9   6_30619745
#> 10   6_4350249

## Parallel computing is supported through the future package
# future::plan(future::multisession, workers = 2)
# multi_lrt(g = g_2, p1 = p1_2, p2 = p2_2)
# future::plan(future::sequential)
# }